services/backend/requirements.txt
Django==4.0.3
django-cors-headers==3.11.0
djangorestframework==3.13.1
psycopg2-binary==2.9.3
gunicorn==20.1.0
Pillow==9.0.1
boto3==1.21.23
django-storages==1.12.3
django-filter==21.1
celery==5.2.7
redis==4.3.4
django-celery-beat==2.3.0
flower==1.2.0
django-celery-results==2.4.0
djangorestframework-api-key==2.2.0
django-nested-admin==3.4.0
pandas==1.5.2
Faker==19.13.0
django-debug-toolbar==4.2.0

# Docs
drf-yasg==1.20.0

# Sentry
sentry-sdk==1.37.1
sentry-sdk [django] == 1.37.1
